LayoutTests/http/tests/security/referrer-policy-header-origin-expected.txt

Tests support for Referrer-Policy HTTP header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing 'Referrer-Policy: origin' - referrer origin: https://127.0.0.1:8443/ - destination origin: https://localhost:8443/ - isMultipartResponse? false
PASS actualReferrer is "https://127.0.0.1:8443/"

Testing 'Referrer-Policy: origin' - referrer origin: https://127.0.0.1:8443/ - destination origin: https://127.0.0.1:8443/ - isMultipartResponse? false
PASS actualReferrer is "https://127.0.0.1:8443/"

Testing 'Referrer-Policy: origin' - referrer origin: https://127.0.0.1:8443/ - destination origin: http://127.0.0.1:8000/ - isMultipartResponse? false
PASS actualReferrer is "https://127.0.0.1:8443/"

Testing 'Referrer-Policy: origin' - referrer origin: https://127.0.0.1:8443/ - destination origin: https://localhost:8443/ - isMultipartResponse? true
PASS actualReferrer is "https://127.0.0.1:8443/"

Testing 'Referrer-Policy: origin' - referrer origin: https://127.0.0.1:8443/ - destination origin: https://127.0.0.1:8443/ - isMultipartResponse? true
PASS actualReferrer is "https://127.0.0.1:8443/"

Testing 'Referrer-Policy: origin' - referrer origin: https://127.0.0.1:8443/ - destination origin: http://127.0.0.1:8000/ - isMultipartResponse? true
PASS actualReferrer is "https://127.0.0.1:8443/"

PASS successfullyParsed is true

TEST COMPLETE
